Refactor Storage Add thread manipulation Add web

+++ src/webapp/storage/FileStorage.java
@@ -13,14 +13,17 @@ import java.util.List;
  * Date: 04.07.2014
  */
 abstract public class FileStorage extends AbstractStorage<File> {
+
     private File dir;
 
     public FileStorage(String path) {
         this.dir = new File(path);
-        if (!dir.isDirectory() || !dir.canWrite()) {
+        if (!dir.isDirectory() || !dir.canWrite())
             throw new IllegalArgumentException("'" + path + "' is not directory or is not writable");
-        }
     }
+
+    protected abstract void doWrite(OutputStream os, Resume resume) throws IOException;
+    protected abstract Resume doRead(InputStream is) throws IOException;
 
     protected void write(File file, Resume resume) {
         try {
@@ -29,10 +32,6 @@ abstract public class FileStorage extends AbstractStorage<File> {
             throw new WebAppException("Couldn't write file " + file.getAbsolutePath(), resume, e);
         }
     }
-
-    protected abstract void doWrite(OutputStream fos, Resume resume) throws IOException;
-
-    protected abstract Resume doRead(InputStream fis) throws IOException;
 
     protected Resume read(File file) {
         try {
@@ -45,20 +44,8 @@ abstract public class FileStorage extends AbstractStorage<File> {
     }
 
     @Override
-    public void doClear() {
-        File[] files = dir.listFiles();
-        if (files == null) return;
-        for (File file : files) {
-            if (!file.delete()) {
-                throw new WebAppException("File " + file.getAbsolutePath() + " cannot be deleted");
-            }
-        }
-    }
-
-    @Override
-    protected boolean exist(String uuid) {
-        File file = getFile(uuid);
-        return file.exists();
+    protected File getCtx(String uuid) {
+        return new File(dir, uuid);
     }
 
     @Override
@@ -67,52 +54,47 @@ abstract public class FileStorage extends AbstractStorage<File> {
     }
 
     @Override
-    protected File getCtx(String uuid) {
-        return getFile(uuid);
+    protected void doClear() {
+        File[] files = dir.listFiles();
+        if (files == null) return;
+        for (File file : files) {
+            if (!file.delete()) {
+                throw new WebAppException("File " + file.getAbsolutePath() + " can not be deleted");
+            }
+        }
     }
 
     @Override
-    public void doSave(File file, Resume r) {
+    protected void doSave(File file, Resume r) {
         try {
             file.createNewFile();
         } catch (IOException e) {
-            throw new WebAppException("Couldn't write file " + file.getAbsolutePath(), r, e);
+            throw new WebAppException("Couldn't create file " + file.getAbsolutePath(), r, e);
         }
         write(file, r);
     }
 
-    private File getFile(String uuid) {
-        return new File(dir, uuid);
-    }
-
     @Override
-    public void doUpdate(Resume r) {
-        File file = getFile(r.getUuid());
+    protected void doUpdate(File file, Resume r) {
         write(file, r);
     }
 
     @Override
-    public Resume doLoad(String uuid) {
-        File file = getFile(uuid);
+    protected Resume doLoad(File file) {
         return read(file);
     }
 
     @Override
-    public void doDelete(String uuid) {
-        File file = getFile(uuid);
-        if (!file.delete()) {
-            throw new WebAppException("File " + file.getAbsolutePath() + " cannot be deleted");
-        }
+    protected void doDelete(File file) {
+        if (!file.delete()) throw new WebAppException("File " + file.getAbsolutePath() + " can not be deleted");
     }
 
     @Override
-    public List<Resume> doGetAll() {
+    protected List<Resume> doGetAll() {
         File[] files = dir.listFiles();
         if (files == null) return Collections.emptyList();
         List<Resume> list = new ArrayList<>(files.length);
-        for (File file : files) {
-            list.add(read(file));
-        }
+        for (File file : files) list.add(read(file));
         return list;
     }
